pyproject.toml
[tool.poetry]
name = "marketing-measurements-op"
version = "0.1.0"
description = "Marketing Measurement Data (Objective Platform)"
authors = ["Matias Cardenas <cardenasmatias.1990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.2.2"
django-ninja = "^1.1.0"
psycopg2 = "^2.9.9"
pytest-django = "^4.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "marketing_op.settings"
addopts = ["--reuse-db", "--no-migrations"]
